remove unused test data
