+++ crates/quote-verifier/data/Intel_SGX_Provisioning_Certification_RootCA.cer
