+++ crates/quote-verifier/data/quote_tdx_00806f050000.dat
