+++ crates/quote-verifier/data/tcb_signing_cert.pem
