+++ crates/quote-verifier/data/tcbinfov3_00806f050000.json
