+++ crates/types/data/quote_v4.dat
