--- data/pck_platform_crl.pem
@@ -1,58 +1,0 @@
------BEGIN X509 CRL-----
-MIIKYjCCCggCAQEwCgYIKoZIzj0EAwIwcDEiMCAGA1UEAwwZSW50ZWwgU0dYIFBD
-SyBQbGF0Zm9ybSBDQTEaMBgGA1UECgwRSW50ZWwgQ29ycG9yYXRpb24xFDASBgNV
-BAcMC1NhbnRhIENsYXJhMQswCQYDVQQIDAJDQTELMAkGA1UEBhMCVVMXDTI0MDUw
-OTA5NDY0MVoXDTI0MDYwODA5NDY0MVowggk0MDMCFG/DTlAj5yiSNDXWGqS4PGGB
-Zq01Fw0yNDA1MDkwOTQ2NDFaMAwwCgYDVR0VBAMKAQEwNAIVAO+ubpcV/KE7h+Mz
-6CYe1tmQqSatFw0yNDA1MDkwOTQ2NDFaMAwwCgYDVR0VBAMKAQEwNAIVAP1ghkhi
-nLpzB4tNSS9LPqdBrQjNFw0yNDA1MDkwOTQ2NDFaMAwwCgYDVR0VBAMKAQEwNAIV
-AIr5JBhOHVr93XPD1joS9ei1c35WFw0yNDA1MDkwOTQ2NDFaMAwwCgYDVR0VBAMK
-AQEwNAIVALEleXjPqczdB1mr+MXKcvrjp4qbFw0yNDA1MDkwOTQ2NDFaMAwwCgYD
-VR0VBAMKAQEwMwIUdP6mFKlyvg4oQ/IFmDWBHthy+bMXDTI0MDUwOTA5NDY0MVow
-DDAKBgNVHRUEAwoBATA0AhUA+cTvVrOrSNV34Qi67fS/iAFCFLkXDTI0MDUwOTA5
-NDY0MVowDDAKBgNVHRUEAwoBATAzAhQHHeB3j55fxPKHjzDWsHyaMOazCxcNMjQw
-NTA5MDk0NjQxWjAMMAoGA1UdFQQDCgEBMDQCFQDN4kJPlyzqlP8jmTf02AwlAp3W
-CxcNMjQwNTA5MDk0NjQxWjAMMAoGA1UdFQQDCgEBMDMCFGwzGeUQm2RQfTzxEyzg
-A0nvUnMZFw0yNDA1MDkwOTQ2NDFaMAwwCgYDVR0VBAMKAQEwNAIVAN8I11a2anSX
-9DtbtYraBNP096k3Fw0yNDA1MDkwOTQ2NDFaMAwwCgYDVR0VBAMKAQEwMwIUKK9I
-W2z2fkCaOdXLWu5FmPeo+nsXDTI0MDUwOTA5NDY0MVowDDAKBgNVHRUEAwoBATA0
-AhUA+4strsCSytqKqbxP8vHCDQNGZowXDTI0MDUwOTA5NDY0MVowDDAKBgNVHRUE
-AwoBATA0AhUAzUhQrFK9zGmmpvBYyLxXu9C1+GQXDTI0MDUwOTA5NDY0MVowDDAK
-BgNVHRUEAwoBATA0AhUAmU3TZm9SdfuAX5XdAr1QyyZ52K0XDTI0MDUwOTA5NDY0
-MVowDDAKBgNVHRUEAwoBATAzAhQHAhNpACUidNkDXu31RXRi+tDvTBcNMjQwNTA5
-MDk0NjQxWjAMMAoGA1UdFQQDCgEBMDMCFGHyv3Pjm04EqifYAb1z0kMZtb+AFw0y
-NDA1MDkwOTQ2NDFaMAwwCgYDVR0VBAMKAQEwMwIUOZK+hRuWkC7/OJWebC7/GwZR
-pLUXDTI0MDUwOTA5NDY0MVowDDAKBgNVHRUEAwoBATAzAhRjnxOaUED9z/GR6KT7
-G/CG7WA5cRcNMjQwNTA5MDk0NjQxWjAMMAoGA1UdFQQDCgEBMDQCFQCVnVM/kknc
-HlE1RM3IML8Zt/HzARcNMjQwNTA5MDk0NjQxWjAMMAoGA1UdFQQDCgEBMDMCFA/a
-Q6ALaOp5t8LerqwLSYvfsq+QFw0yNDA1MDkwOTQ2NDFaMAwwCgYDVR0VBAMKAQEw
-NAIVAJ1ndTuB5HCQrqdj++xMRUm825kzFw0yNDA1MDkwOTQ2NDFaMAwwCgYDVR0V
-BAMKAQEwMwIUNL+7eh2cVoFH4Ri2FPe3btPvaN8XDTI0MDUwOTA5NDY0MVowDDAK
-BgNVHRUEAwoBATA0AhUAhdPJOBt3p+BNEZyeWtZ0n/P/q4cXDTI0MDUwOTA5NDY0
-MVowDDAKBgNVHRUEAwoBATA0AhUAk4h8pEEeepI70f7SgZspSfIBtbQXDTI0MDUw
-OTA5NDY0MVowDDAKBgNVHRUEAwoBATAzAhQkmNxig5MJlv2L8jo3rL4mo77UVxcN
-MjQwNTA5MDk0NjQxWjAMMAoGA1UdFQQDCgEBMDQCFQCKZvGnSUiGZ2icw5A6xUxm
-K3EucxcNMjQwNTA5MDk0NjQxWjAMMAoGA1UdFQQDCgEBMDQCFQCvwTYQvdNst5hd
-EGSBqIDToB/aBxcNMjQwNTA5MDk0NjQxWjAMMAoGA1UdFQQDCgEBMDQCFQDv4Ess
-M9A2qslspnO/HppHtk1cuxcNMjQwNTA5MDk0NjQxWjAMMAoGA1UdFQQDCgEBMDQC
-FQCD2ayNi7UJ0cbICa1xLoQwVZ7X8xcNMjQwNTA5MDk0NjQxWjAMMAoGA1UdFQQD
-CgEBMDMCFHkx/VC1Bxwbv8W3tt7YtFudi4UpFw0yNDA1MDkwOTQ2NDFaMAwwCgYD
-VR0VBAMKAQEwMwIUH6IOKXC95dV/e43fgzlITh8dCCMXDTI0MDUwOTA5NDY0MVow
-DDAKBgNVHRUEAwoBATAzAhQeh7LDsy2NI+QRzvNBl7la8Mit9RcNMjQwNTA5MDk0
-NjQxWjAMMAoGA1UdFQQDCgEBMDQCFQCa/S7pCkc1UKFn2ZaRFDfHUC0fCRcNMjQw
-NTA5MDk0NjQxWjAMMAoGA1UdFQQDCgEBMDMCFESBsPEXKKE7aW0+qcdwoLFexY3a
-Fw0yNDA1MDkwOTQ2NDFaMAwwCgYDVR0VBAMKAQEwNAIVAKeFn1eYLvDmfTe8jvLv
-Wsg1/xqpFw0yNDA1MDkwOTQ2NDFaMAwwCgYDVR0VBAMKAQEwMwIUeuN3SKn5EvTG
-O6erB8WTzh0dEYEXDTI0MDUwOTA5NDY0MVowDDAKBgNVHRUEAwoBATAzAhQTiEsz
-Jpk4wZWqFw/KddoXdTjfCxcNMjQwNTA5MDk0NjQxWjAMMAoGA1UdFQQDCgEBMDMC
-FCw8xv6SedsVFtXOOfKomM2loXXhFw0yNDA1MDkwOTQ2NDFaMAwwCgYDVR0VBAMK
-AQEwMwIUcXlIaHUJI0vpeeS33ObzG+9ktowXDTI0MDUwOTA5NDY0MVowDDAKBgNV
-HRUEAwoBATA0AhUAnXbvLDnBNuhli25zlrHXRFonYx8XDTI0MDUwOTA5NDY0MVow
-DDAKBgNVHRUEAwoBATA0AhUAw+Al/KmV829ZtIRnk54+NOY2Gm8XDTI0MDUwOTA5
-NDY0MVowDDAKBgNVHRUEAwoBATA0AhUAjF9rMlfaBbF0KeLmG6ll1nMwYGoXDTI0
-MDUwOTA5NDY0MVowDDAKBgNVHRUEAwoBATA0AhUAoXxRci7B4MMnj+i98FIFnL7E
-5kgXDTI0MDUwOTA5NDY0MVowDDAKBgNVHRUEAwoBAaAvMC0wCgYDVR0UBAMCAQEw
-HwYDVR0jBBgwFoAUlW9dzb0b4elAScnU9DPOAVcL3lQwCgYIKoZIzj0EAwIDSAAw
-RQIhAJf932bqsDTVzc+H/DBN2bICwkKeY418cqKack0N0vHJAiAzhNcINTRHNKt5
-ZZNbPPSAdrL5ACMgO5HsYn9sF2w7Jw==
------END X509 CRL-----
