--- data/qeidentityv2_apiv4.json
@@ -1,1 +1,0 @@
-{"enclaveIdentity":{"id":"TD_QE","version":2,"issueDate":"2024-09-10T06:27:30Z","nextUpdate":"2024-10-10T06:27:30Z","tcbEvaluationDataNumber":16,"miscselect":"00000000","miscselectMask":"FFFFFFFF","attributes":"11000000000000000000000000000000","attributesMask":"FBFFFFFFFFFFFFFF0000000000000000","mrsigner":"DC9E2A7C6F948F17474E34A7FC43ED030F7C1563F1BABDDF6340C82E0E54A8C5","isvprodid":2,"tcbLevels":[{"tcb":{"isvsvn":4},"tcbDate":"2023-08-09T00:00:00Z","tcbStatus":"UpToDate"}]},"signature":"4994904c51f5a425e32ec0e8a79e8ae4b5510fe462b262a21c34e50d241a621f6cb7ebabb7876a3b97b11a34c3ca41c32deb25e3b1d0e38581477c177013e652"}
